Week 14: 半休（0.5日）対応実装完了 ✅
【実装内容】
- vacation_model.dart: VacationAccrual.daysGranted を int → double
- vacation_repository.dart: recordVacationAccrual() パラメータを int → double
- vacation_repository.dart: getAccrualSummary() で fold<double>
- vacation_repository.dart: VacationSummary.totalGranted を int → double
- vacation_settings_screen.dart: item.daysGranted.toDouble()
- database_helper.dart: DB v15 マイグレーション（is_carried_over カラム追加）
- shift_management_screen.dart: _patterns に「有休」「半休」パターン追加
- shift_management_screen.dart: loadShifts() で widget.patterns → _patterns に修正

【テスト結果】
✅ 「有休」「半休」がカレンダーに正しく表示される
✅ VacationUsage に 1.0日 / 0.5日 が正確に記録される
✅ SQLiteLog エラーが消えた

【未解決（Week 15で対応）】
❌ 有給残日数が減らない
❌ カレンダーの日付削除機能

+++ lib/repositories/vacation_repository.dart
@@ -101,7 +101,7 @@ class VacationRepository {
   Future<void> recordVacationAccrual(
     String userId,
     DateTime accrualDate,
-    int daysGranted, {
+    double daysGranted, {
     String? notes,
   }) async {
     final db = await _dbHelper.database;
@@ -363,7 +363,7 @@ class VacationRepository {
     final accruals = await getVacationAccrualsByUser(userId);
     if (accruals.isEmpty) return null;
 
-    final totalGranted = accruals.fold<int>(
+    final totalGranted = accruals.fold<double>(
       0,
       (sum, accrual) => sum + accrual.daysGranted,
     );
@@ -531,7 +531,7 @@ class VacationRepository {
 
 /// 付与サマリー（UI表示用）
 class VacationSummary {
-  final int totalGranted;        // 累計付与日数
+  final double totalGranted;        // 累計付与日数
   final double totalUsed;        // 累計使用日数
   final double totalRemaining;   // 累計残日数
   final int expiringCount;       // 失効予定数
